requirements.txt
asttokens==3.0.1
attrs==25.4.0
branca==0.8.2
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
fastjsonschema==2.21.2
folium==0.20.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.15.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
platformdirs==4.5.1
plotly==6.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sodapy==2.2.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
xyzservices==2025.11.0
plotly==6.5.1
